refactor(inspector): material inspector is now a normal window
