+++ editor/src/DocumentWindows/InspectorWindow.cpp
@@ -41,63 +41,10 @@ extern ImGuiID g_materialInspectorDockID;
 
 namespace nexo::editor
 {
-	void MaterialInspector::setup()
-	{
-		renderer::FramebufferSpecs framebufferSpecs;
-		framebufferSpecs.attachments = {renderer::FrameBufferTextureFormats::RGBA8, renderer::FrameBufferTextureFormats::Depth};
-		framebufferSpecs.width = static_cast<unsigned int>(64);
-		framebufferSpecs.height = static_cast<unsigned int>(64);
-		m_framebuffer = renderer::Framebuffer::create(framebufferSpecs);
-		m_framebuffer->setClearColor({0.05f, 0.05f, 0.05f, 0.0f});
-	}
-
-	void MaterialInspector::show(const int selectedEntity)
-	{
-		static bool materialModified = true;
-		static utils::ScenePreviewOut previewParams;
-
-		if (selectedEntity != -1)
-		{
-			if (m_ecsEntity != selectedEntity)
-				m_ecsEntity = selectedEntity;
-   		}
-
-		if (m_ecsEntity == -1)
-			return;
-
-		if (materialModified)
-		{
-			utils::genScenePreview("Modify material inspector", {64, 64}, m_ecsEntity, previewParams);
-			auto &app = nexo::getApp();
-			auto &cameraComponent = nexo::Application::m_coordinator->getComponent<components::CameraComponent>(previewParams.cameraId);
-			cameraComponent.clearColor = {0.05f, 0.05f, 0.05f, 0.0f};
-			app.run(previewParams.sceneId, RenderingType::FRAMEBUFFER);
-			m_framebuffer = cameraComponent.m_renderTarget;
-			materialModified = false;
-			app.getSceneManager().deleteScene(previewParams.sceneId);
-		}
-
-		// --- Material preview ---
-		if (m_framebuffer->getColorAttachmentId(0) != 0)
-			ImGui::Image(static_cast<ImTextureID>(static_cast<intptr_t>(m_framebuffer->getColorAttachmentId(0))), {64, 64}, ImVec2(0, 1), ImVec2(1, 0));
-		ImGui::SameLine();
-
-		auto materialVariant = m_inspector.getSubInspectorData<MaterialInspector>();
-		if (std::holds_alternative<components::Material*>(materialVariant)) {
-		    auto materialPtr = std::get<components::Material*>(materialVariant);
-			materialModified = Widgets::drawMaterialInspector(materialPtr);
-		}
-	}
-
     InspectorWindow::~InspectorWindow() = default;
 
     void InspectorWindow::setup()
     {
-
-    	m_materialInspector = std::make_shared<MaterialInspector>(*this);
-    	m_materialInspector->setup();
-
-
 		m_entityProperties[typeid(components::TransformComponent)] = std::make_shared<TransformProperty>(*this);
 		m_entityProperties[typeid(components::RenderComponent)] = std::make_shared<RenderProperty>(*this);
 		m_entityProperties[typeid(components::AmbientLightComponent)] = std::make_shared<AmbientLightProperty>(*this);
@@ -115,6 +62,8 @@ namespace nexo::editor
     void InspectorWindow::show()
     {
         ImGui::Begin("Inspector", &m_opened, ImGuiWindowFlags_NoCollapse);
+       	if (m_firstOpened)
+        	firstDockSetup("Inspector");
         auto const &selector = Selector::get();
         const int selectedEntity = selector.getSelectedEntity();
 
@@ -131,31 +80,6 @@ namespace nexo::editor
         }
 
         ImGui::End();
-
-        if (getSubInspectorVisibility<MaterialInspector>())
-        {
-        	static bool first = true;
-
-            ImGuiWindowFlags window_flags = ImGuiWindowFlags_NoCollapse;
-            if (first)
-            	window_flags |= ImGuiWindowFlags_NoBringToFrontOnFocus;
-
-            if (ImGui::Begin("Material Inspector", &getSubInspectorVisibility<MaterialInspector>(), window_flags))
-            {
-	            if (ImGuiWindow* currentWindow = ImGui::GetCurrentWindow(); currentWindow && first)
-	            {
-		            const bool isDocked = currentWindow->DockIsActive;
-	                const ImGuiID currentDockID = currentWindow->DockId;
-
-					//TODO: Implement a docking registry
-	                if (!isDocked || currentDockID != g_materialInspectorDockID)
-						currentWindow->DockId = g_materialInspectorDockID;
-	            }
-                m_materialInspector->show(selectedEntity);
-                first = false;
-            }
-            ImGui::End();
-        }
     }
 
     void InspectorWindow::showSceneProperties(const scene::SceneId sceneId)
