+++ editor/src/DocumentWindows/MaterialInspector.cpp
@@ -1,0 +1,102 @@
+//// MaterialInspector.cpp ///////////////////////////////////////////////////////////////
+//
+//  zzzzz       zzz  zzzzzzzzzzzzz    zzzz      zzzz       zzzzzz  zzzzz
+//  zzzzzzz     zzz  zzzz                    zzzz       zzzz           zzzz
+//  zzz   zzz   zzz  zzzzzzzzzzzzz         zzzz        zzzz             zzz
+//  zzz    zzz  zzz  z                  zzzz  zzzz      zzzz           zzzz
+//  zzz         zzz  zzzzzzzzzzzzz    zzzz       zzz      zzzzzzz  zzzzz
+//
+//  Author:      Mehdy MORVAN
+//  Date:        25/03/2025
+//  Description: Source file for the material inspector window
+//
+///////////////////////////////////////////////////////////////////////////////
+
+#include "MaterialInspector.hpp"
+#include "DocumentWindows/InspectorWindow.hpp"
+#include "utils/ScenePreview.hpp"
+#include "components/Camera.hpp"
+#include "Components/Widgets.hpp"
+#include "context/Selector.hpp"
+
+namespace nexo::editor {
+	void MaterialInspector::setup()
+	{
+		renderer::FramebufferSpecs framebufferSpecs;
+		framebufferSpecs.attachments = {renderer::FrameBufferTextureFormats::RGBA8, renderer::FrameBufferTextureFormats::Depth};
+		framebufferSpecs.width = static_cast<unsigned int>(64);
+		framebufferSpecs.height = static_cast<unsigned int>(64);
+		m_framebuffer = renderer::Framebuffer::create(framebufferSpecs);
+		m_framebuffer->setClearColor({0.05f, 0.05f, 0.05f, 0.0f});
+	}
+
+	void MaterialInspector::shutdown()
+	{
+
+	}
+
+	void MaterialInspector::renderMaterialInspector(int selectedEntity)
+	{
+		static bool materialModified = true;
+		static utils::ScenePreviewOut previewParams;
+
+		if (selectedEntity != -1)
+		{
+			if (m_ecsEntity != selectedEntity)
+				m_ecsEntity = selectedEntity;
+   		}
+
+		if (m_ecsEntity == -1)
+			return;
+
+		if (materialModified)
+		{
+			utils::genScenePreview("Modify material inspector", {64, 64}, m_ecsEntity, previewParams);
+			auto &app = nexo::getApp();
+			auto &cameraComponent = nexo::Application::m_coordinator->getComponent<components::CameraComponent>(previewParams.cameraId);
+			cameraComponent.clearColor = {0.05f, 0.05f, 0.05f, 0.0f};
+			app.run(previewParams.sceneId, RenderingType::FRAMEBUFFER);
+			m_framebuffer = cameraComponent.m_renderTarget;
+			materialModified = false;
+			app.getSceneManager().deleteScene(previewParams.sceneId);
+		}
+
+		// --- Material preview ---
+		if (m_framebuffer->getColorAttachmentId(0) != 0)
+			ImGui::Image(static_cast<ImTextureID>(static_cast<intptr_t>(m_framebuffer->getColorAttachmentId(0))), {64, 64}, ImVec2(0, 1), ImVec2(1, 0));
+		ImGui::SameLine();
+
+		auto inspectorWindow = m_windowRegistry.getWindow<InspectorWindow>();
+		auto materialVariant = inspectorWindow->getSubInspectorData<MaterialInspector>();
+		if (std::holds_alternative<components::Material*>(materialVariant)) {
+		    auto materialPtr = std::get<components::Material*>(materialVariant);
+			materialModified = Widgets::drawMaterialInspector(materialPtr);
+		}
+	}
+
+	void MaterialInspector::show()
+	{
+  		auto const &selector = Selector::get();
+        const int selectedEntity = selector.getSelectedEntity();
+        auto inspectorWindow = m_windowRegistry.getWindow<InspectorWindow>();
+		if (inspectorWindow->getSubInspectorVisibility<MaterialInspector>())
+        {
+            ImGuiWindowFlags window_flags = ImGuiWindowFlags_NoCollapse;
+            if (m_firstOpened)
+            	window_flags |= ImGuiWindowFlags_NoBringToFrontOnFocus;
+
+            if (ImGui::Begin("Material Inspector", &inspectorWindow->getSubInspectorVisibility<MaterialInspector>(), window_flags))
+            {
+            	if (m_firstOpened)
+             		firstDockSetup("Material Inspector");
+                renderMaterialInspector(selectedEntity);
+            }
+            ImGui::End();
+        }
+	}
+
+	void MaterialInspector::update()
+	{
+
+	}
+}
